# src/OneHope.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OneHope.API",
    "description": "Esta API provee servicios para alquilar, comprar, devolver o reabastecer portátiles.",
    "contact": {
      "name": "OneHope Team",
      "email": "daniel.tomas@alu.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Portatiles/GetPortatilesParaPedido": {
      "get": {
        "tags": [
          "Portatiles"
        ],
        "operationId": "GetPortatilesParaPedido",
        "parameters": [
          {
            "name": "filtroModelo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroMarca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroStockMinimo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "filtroStockMaximo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "filtroProveedor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroNombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaPedidoDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaPedidoDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaPedidoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Portatiles/GetPortatilesParaDevolver": {
      "get": {
        "tags": [
          "Portatiles"
        ],
        "operationId": "GetPortatilesParaDevolver",
        "parameters": [
          {
            "name": "idCompra",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "CustomerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "precio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilesParaDevolverDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilesParaDevolverDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilesParaDevolverDTO"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/Portatiles/GetPortatilesParaComprar": {
      "get": {
        "tags": [
          "Portatiles"
        ],
        "operationId": "GetPortatilesParaComprar",
        "parameters": [
          {
            "name": "nombrePortatil",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modeloPortatil",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "marcaPortatil",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "procesadorPortatil",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ramPortatil",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "precioPortatil",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "stock",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaComprarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaComprarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaComprarDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Portatiles/GetPortatilesParaAlquiler": {
      "get": {
        "tags": [
          "Portatiles"
        ],
        "operationId": "GetPortatilesParaAlquiler",
        "parameters": [
          {
            "name": "filtroMarca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroProcesador",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroRam",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroModelo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaAlquilerDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaAlquilerDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortatilParaAlquilerDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PortatilParaAlquilerDTO": {
        "required": [
          "id",
          "marca",
          "modelo",
          "precioAlquiler",
          "procesador",
          "ram",
          "stockAlquilar"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "modelo": {
            "minLength": 1,
            "type": "string"
          },
          "marca": {
            "minLength": 1,
            "type": "string"
          },
          "stockAlquilar": {
            "type": "integer",
            "format": "int32"
          },
          "precioAlquiler": {
            "type": "number",
            "format": "currency"
          },
          "procesador": {
            "minLength": 1,
            "type": "string"
          },
          "ram": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PortatilParaComprarDTO": {
        "required": [
          "marca",
          "modelo",
          "nombre",
          "precioCompra",
          "procesador",
          "ram",
          "stock"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "modelo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "precioCompra": {
            "minimum": 1,
            "type": "number",
            "format": "currency"
          },
          "ram": {
            "minLength": 1,
            "type": "string"
          },
          "marca": {
            "minLength": 1,
            "type": "string"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "procesador": {
            "minLength": 1,
            "type": "string"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PortatilParaPedidoDTO": {
        "required": [
          "id",
          "marca",
          "modelo",
          "precioCoste",
          "proveedor",
          "stock"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "modelo": {
            "minLength": 1,
            "type": "string"
          },
          "marca": {
            "minLength": 1,
            "type": "string"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "precioCoste": {
            "type": "number",
            "format": "currency"
          },
          "proveedor": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PortatilesParaDevolverDTO": {
        "required": [
          "Cantidad",
          "Marca",
          "Modelo",
          "Total"
        ],
        "type": "object",
        "properties": {
          "IdCompra": {
            "type": "integer",
            "format": "int32"
          },
          "IdPortatil": {
            "type": "integer",
            "format": "int32"
          },
          "IdLineaCompra": {
            "type": "integer",
            "format": "int32"
          },
          "Marca": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "Modelo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "Cantidad": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "FechaCompra": {
            "type": "string",
            "format": "date"
          },
          "Total": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}